python workers should die on ctrl-c (SIG_INT)
if you write a python worker and run it in the foreground, you
couldn't kill it with control-c which is *super* annoying during
development.

This is because python adds a default signal handler that must be
handled by the main thread.  When a gearbox worker calls run() it
blocks the main thread in C++ land.  The fix is to tell python to
ignore SIG_INT so C++ can handle it