change install location of sql_status plugin
sql_status.so was being installed to /usr/lib64.  First off,
StatusManager defaults to looking in /usr/libexec/gearbox/plugins/status/
(although that is configurable at build time).   You can override where
StatusManager looks at runtime by setting status.plugindir, but if you set that
to /usr/lib64 it will try to load all the .so files in that directory and you'll
end up with an error like this:

    failed to load plugin: /usr/lib64/libthread_db.so: undefined symbol:
    ps_lgetfpregs

A better solution is to modify make install (and the rpm) so it installs
it where the statusmanager is looking for it.