rpm packaging: sql-status-plugin requires soci driver
My /etc/gearbox/config.d/status.conf specified that db_type was
sqlite.  Without soci-sqlite3 installed I was getting this error
in /var/log/gearbox/gearbox.log:

    Failed to find shared library for backend sqlite3

The solution was to install the needed soci library.  It's possible
that you could argue that the dependency should be on the same package
that owns /etc/gearbox/config.d/status.conf since it is what determines
which soci driver we need, but this should cover the common use cases