perl unit tests should use locally built perl-lib
If you have the perl swig libraries installed in somewhere in your perl
path they would get used instead of the locally built ones in your
working directory.

There was an attempt to do this, but it stuck the local directories at
the end of the perl library search path, instead of the beginning.  This
commit fixes that.