python swig string to uri typemap can segfault
The python chained test worker (not yet committed) was segfaulting when
calling the job->base_uri method.

The string to uri typemap was using PyString_Size to get a length to
pass to std::string assign method.  PyString_Size returns a Py_ssize_t
which is usually a typedef for ssize_t  See Python's PEP 353
http://www.python.org/dev/peps/pep-0353/

The signature for std::string's assign is
    string& assign (const char* s, size_t n);

We were blinding converting a ssize_t to a size_t.  It turns out that
that was actually totally unnecessary because PyString_AsString already
returns a NUL terminated string, so we can use the assign method with
this signature:
    string& assign (const char* s);