python "error return without exception set"
When a C++ method threw a Gearbox::Error exception, the swig
exception handler is supposed to map it to a python gearbox
Error class.  That wasn't happening because the class couldn't
be found because the module name wasn't being passed to
PyErr_NewException.  Instead, the python error indicator wasn't
being set and the function wasn't being returned.  This resulted
in the very unhelpful message "error return without exception set"
so the user had no idea what they did to cause the exception.